fix(psd): Improve memory efficiency of PSD read
The PSD reader was written 12 years ago by a GSOC student and has
changed little since. The following is not a knock on that student,
who clearly did a good job if his code has survived mostly unomdified
for over a decade. In chasing down problems with handling extremely
large psd files, I've noticed some oddities and inefficiencies that
need fixing. It's not an emergency, so I'll handle it in pieces.

The item addressed by this patch is memory usage and thread
concurrency. It turns out that the reader allocates memory for the
scanline reads as it goes... and basically keeps it in the ImageInput
until it's destroyed, even though it will never actually be used
again. I guess it's just trying to save the next allocation if it
reads the same scanline again or on the next subimage? But this means
that if you read a really big psd image, it's basically doubling the
memory use -- it reads one copy into the user's buffer, but also holds
on to this extra copy for no reason. Also, a bunch of the decoding
helper functions implicitly use this memory, which means we lock a
mutex for the duration of read_native_scanline (which in theory should
not be necessary since all I/O is happening through thread-safe
IOProxy calls).

So this patch does the following:

1. Remove both this retained scanline memory as well as some other
   scratch memory from the ImageInput, and make it allocated and freed
   local to read_native_scanline.

2. Change all the helper methods that used it implicitly so that they
   are passed buffers to use. In many cases these have been changed to
   `const` methods and in a few cases to `static`, making it much
   easier to assure ourselves that they don't modify (or sometimes use
   at all) anything in the PSDInput to do their task.

3. The result of these changes means that read_native_scanline doen't
   need to modify any of the data methods of the class (though it does
   use, without modifcation, some things set up in the course of
   open).  Because it's now non-mutable, we can also remove the mutex,
   now allowing multiple threads to read scanlines from the same file
   simultaneously without blocking each other.

4. Some other miscellaneous cleanup, including that some scratch
   memory was allocated as std::string (!), whereas it should really
   be a `vector<unsigned char>` or `unique_ptr<unsigned char>`, and
   also changing some computations from double to float.

There is still some more work to do after this is merged. In
particular, I think now that scanline read is stateless / non-mutable
with respect to the PSDInput object, this is a candidate for adding a
multi-scanline read_native_scanlines method (replacing the default
implementation of looping and calling read_native_scanline serially
for each scanline), which should also be able to read multiple
scanlines with some overhead and allocations done once per block, and
with blocks of scanlines running in parallel using the thread pool.
This should be a huge efficiency gain when reading psd files. But it
will wait until the next patch.