build(cmake): Fix incorrect variable name to control symbol visibility
As Fabrice Macagno points out on the mail list, we weren't setting the
correct variable name CMAKE_CXX_VISIBILITY_PRESET.

Fixing this -- and finally hiding all the symbols correctly --
uncovered an instance where testtex was relying on some classes that
should have been hidden in libOpenImageIO. For symbol visibility
reasons, move the functionality of testtex's test_hash() function to
be in TextureSystem itself. This is a much cleaner separation.