build(deps): fmt 10.0 support
fmt 10.0 was recently released, and it breaks a lot for us because
they have fully removed legacy support for automatically being able to
format anything that has an iostream `<<` operator. This facility was
deprecatd in 8.1 but could be re-enabled with FMT_DEPRECATED_OSTREAM,
but even that was removed in 10.0.

So this means that we had to write several explicit format handlers
for some types for use with format() and print().

Also, they are even more strict with sprintf/printf which don't honor
the custom formatters as far as I can tell, so in several cases I also
switched those to the std::format/print style.