fix(oiiotool): don't propagate unsupported channels
Fixes 3837

oiiotool already attempted to trim extra channels when outputting to a
format that can't support it. It correctly handled the two common
cases:

* More than 4 channels for formats that don't have
  `supports("nchannels")` will drop after the first 4 channels.
* More than 3 channels for formats that don't have `supports("alpha")`
  will drop after the first 3 channels.

But this omits an interesting (but rare) corner case: a 4 channel file
where the last channel is NOT alpha, and the output file supports an
alpha channel but not arbitrary channels. A concrete example might be
an input file that is exr with channels R, G, B, Z, writing to a PNG
file (which can support RGB or RGBA). We wouldn't want to write what
appears to be an RGBA file if that last channel isn't really alpha.

Also, noticed that oiiotool --chnames didn't properly clear the
alpha_channel and z_channel flags, which is important if the new
channel lists drop one of those channels.