ci: Get icc tests working again
For a couple years, we installed the
intel-oneapi-compiler-dpcpp-cpp-and-cpp-classic package from Intel's
yum repo https://yum.repos.intel.com/oneapi to get a copy of icc and
icx to test against.

This broke in January 2023 when Intel bumped the version of the
package and the new package's icc (but curiously not icx) wouldn't run
on the CentOS images and complained about

    /opt/intel/oneapi/compiler/2023.1.0/linux/bin/intel64/icpc: /lib64/libstdc++.so.6: version `GLIBCXX_3.4.21' not found (required by /opt/intel/oneapi/compiler/2023.1.0/linux/bin/intel64/icpc)

In https://github.com/OpenImageIO/oiio/pull/3744, we tried
to combat this by locking down to a specific version of icc 2022.1
that was known to work and that we used to use until the default
version was bumped.

But in April 2023, the 2022.1.0-3768 package seemed to disappear from
the repository entirely, so we're back to being unable to install icc
and test against it from any of the ASWF VFX Platform compliant
containers. So in this patch, we disable that part of the CI testing.

However, with some tinkering, I was able to use that icc package from
the plain Ubuntu 22.04 environment (just not from the VFX Platform
compliant container). So set that up so we're still testing icc
somehow.

icx has been fine all along, this doesn't change anything about how
we test icx, which is still done in the VFX Platform 2022 container.

Signed-off-by: Larry Gritz <lg@larrygritz.com>