admin: Add draft ASWF documents to the repo
Establish an ASWF subdirectory.

Add draft or placeholder CLAs, CoC, governance (including TSC member
list), and project intake checklist.

Suggestions, amendments, and additional TSC members should be
communicated via review comments on this PR (prior to merge) or
further PRs against these documents (after the merge, while they are
still in draft / unvoted form),

The CLAs and CoC are presumed final, since they are merely copies of
the standard ones for ASWF. Project intake checklist replicated from
OSL, will be replaced when a new one is given to us (but is presumably
going to be very similar, so this will help us get started).
Governance document is copied from OSL, and is for the OIIO TSC to
review and vote on to arrive at a final form appropriate for this
project. The technical charter is currently an empty placeholder, as
we are still resolving questions about the proposed draft with LF.

Signed-off-by: Larry Gritz <lg@larrygritz.com>