fix(exr): try to improve exr thread pool weirdness
There's some code I put in not long ago to automatically call exr's
thread pool setNumThreads(0) upon exit, as an attempt to work around
some exr thread pool deadlock conditions. But... we've seen some
weirdness from this workaround as well, so try to limit its
application by not bothering when using OpenEXR >= 3.1.8 (at which
point we believe the exr side bug was fixed) and also to do so only on
Windows.

Signed-off-by: Larry Gritz <lg@larrygritz.com>