fix(psd): Prevent simultaneous psd thumbnail reads from clashing
When reading a PSD file, if it contains a thumbnail, we read it from
the in-memory blob with a combination of an IOMemReader and an
ImageBuf.  But... we always named it "thumbnail.jpg" without
considering that multiple simultaneous PSD files reading identically
named images (but which are in fact different) could cause a weird
kind of clashing in any underlying use of ImageCache!

The simple fix is to use ordinary ImageInput approach to reading the
thumbnail from the IOMemReader, but just store it in the ImageBuf.

Fixes 3824

Signed-off-by: Larry Gritz <lg@larrygritz.com>