Fix: Wait for terminated threads to join in thread_pool::Impl::resize
Part of solving #3851.

When decreasing the number of threads in `thread_pool::Impl::resize`
it was possible for the threads marked for termination to still be
fully alive at process exit time if the call to `resize` and the exit
of the main function were close enough to each other.

This change makes `resize` wait for the threads to actually join
guaranteeing the threads have been shut down in a orderly manner
when the `resize` function exits.