Fix: Add shutdown method
This change adds a `shutdown` method to the API that can be used to perform
shutdown of the thread-pool(s) (or other clean-up activities). Applications
that utilize OpenImageIO must call it before exiting their main function to
prevent possible deadlocks during application shutdown (see #3851)

This change also updates the documentation and all oiio-tools to call the
new `shutdown` method.

This is the final change for fixing #3851