fix(exr): correction to dwa vs zip logic when outputting OpenEXR
PR #2147, addressing issue #1844, is related to an apparent OpenEXR
bug wherein single channel tiled images where the tile dimensions were
not a power of 2 and >= 16 would not work properly with dwa
compression. We work around the problem by automatically substituting
zip compression in those cases. But it was implemented with a subtle
bug, forgetting that tile_size == 0 for untiled images, and thus we
were silently switching from dwa to zip compression for all *untiled*
single channel images of any size (which presumably were never
problematic). This patch fixes the test so we are not changing the
compression for the untiled files.

There's a long

Signed-off-by: Larry Gritz <lg@larrygritz.com>