dev(ustring.h): string literal operator for ustring and ustringhash
* Add string literal operators for ustring and ustringhash so you can
  write:
  ```
      "foo"_us
      "bar"_ush
  ```
  Note that on Cuda (but not on the host), this ustringhash construction
  is constexpr.

* Add a new `ustringhash(const char*, size_t)` constructor, which is also
  constexpr on the device side.

* Mark the existing `ustringhash(string_view)` constructor as constexpr
  on the device side.

We can mark those things as constexpr on the device side because on
the Cuda side, the implementation only takes the hash (which is
constexpr), and does not actually add a ustring to the table (which
cannot be constexpr). To make such declarations easy, add a new macro
to platform.h, `OIIO_DEVICE_CONSTEXPR` which is constexpr for the
device, merely inline for the host.

Signed-off-by: Larry Gritz <lg@larrygritz.com>