exr: Re-enable openexr core default when recent enough
We had disabled the use of the "openexr core" library by default,
even for OpenEXR versions that nominally support it, when it was
discovered that it stil didn't support the DWA compression types.

That has finally been added to the OpenEXR trunk as well as the
OpenEXR 3.1.10 release. When we are building against a sufficiently new
OpenEXR to fully support all the compression types for exr core, then
re-enable it by default.

Bump the "latest" CI tests to OpenEXR 3.1.10.

Signed-off-by: Larry Gritz <lg@larrygritz.com>