feat(api): ImageInput::open_check
Not unlike the ImageOutput::open_check we added a few months back,
this creates an ImageInput utility function that centralizes a number
of validation checks. Currently, this is primarily a check on allowed
resolution and total size, aimed at making a guess at files that are
corrupted or malicious.

Also adds a new `supports("noimage")` query.

Signed-off-by: Larry Gritz <lg@larrygritz.com>