docs: Prototype docs examples doubling as tests
New theory of code examples in the docs: every one should live in the
testsuite somewhere, and merely be included by reference in the docs.
This ensures that every code example works, stays current with the
evolution of the APIs, and never breaks. It also beefs up our test
cases for more thorough code overage in the testsuite.

I've done this with just two examples here, but hopefully it serves as
an example for others to -- bit by bit -- convert all the other code
examples in the docs into tests in a similar manner.

Doing this for any one example is a perfect "good first issue," since
it's bite sized and can easily be done in a day, it doesn't require
any knowledge of deep OIIO internals, and in the process it also
teaches you something about how both the docs and the testsuite are
set up.

A few notes:

* I'm structuring it (for now?) as two testsuite entry for each
  chapter of the documentation (one for C++ and one for Python).

* Note how I can have many code snippets in a test source file, and
  use the Sphinx `:start-after:` and `:end-before:` modifiers of
  `literalinclude` to clip particular subsections of the file, using
  special comment markers in the code.

While I was there, I also fixed some typos I noticed in the text of
the chapter, oops.

Signed-off-by: Larry Gritz <lg@larrygritz.com>