perf(oiiotool): oiiotool --mosaic should avoid unnecessary type conversion
`oiiotool --mosaic` was always allocating a `float` buffer for the
result.  Instead, make it use the "widest" data type of the input
images. So, for example, if you are making a mosaic of uint8 images,
the result buffer should be uint8, not have to allocate 4x bigger and
do uint8->float conversions as it pastes the images.

Signed-off-by: Larry Gritz <lg@larrygritz.com>