build: Test ABI compliance
This patch adds a new test to the matrix: verifying that we do not
break the ABI during releases.

Basically, in addition to the current checkout, it also builds a
checkout of the designated "ABI standard" (generally, the first
non-beta release of a major version), then generates ABI dumps of both
versions and runs the ABI compliance checker using the tools from
https://github.com/lvc

Making two builds sounds expensive, but both builds are rigged to be
minimal, since only the main two libraries are needed for ABI checks
(we can turn off the binary tools, unit tests, python bindings, and
support for many formats), and of course the test suite need not run.
So actually, it's less expensive than most of the test matrix entries,
and it helps ensure that we don't accidentally break ABI when we don't
mean to.

Signed-off-by: Larry Gritz <lg@larrygritz.com>